feat: add validation and error instances and status codes

+++ routes/client.js
@@ -1,5 +1,7 @@
 const express = require('express')
+
 const { db } = require('../utils/db')
+const { NotFoundError } = require('../utils/errors')
 
 const clientRouter = express.Router()
 
@@ -13,20 +15,31 @@ clientRouter
         res.render('client/add-client')
     })
     .get('/edit/:id', (req, res) => {
+        const client = db.getOne(req.params.id)
+
+        if(!client){
+            throw new NotFoundError()
+        }
+
         res.render('client/edit-client', {
             client: db.getOne(req.params.id)
         })
     })
     .get('/:id', (req, res) => {
+        const client = db.getOne(req.params.id)
+
+        if(!client){
+            throw new NotFoundError()
+        }
+
         res.render('client/one-client', {
-            client: db.getOne(req.params.id)
+            client
         })
     })
     .post('/', async (req, res, next) => {
         try {
            const id = await db.create(req.body)
-
-            res.render('client/added',{
+            res.status(201).render('client/added',{
                 id
             })
         } catch(error){
@@ -34,12 +47,24 @@ clientRouter
         }
     })
     .put('/:id', (req, res) => {
+        const client = db.getOne(req.params.id)
+
+        if(!client){
+            throw new NotFoundError()
+        }
+
         db.update(req.params.id, req.body)
         res.render('client/updated', {
             id: req.params.id
         })
     })
     .delete('/:id', (req, res) => {
+        const client = db.getOne(req.params.id)
+
+        if(!client){
+            throw new NotFoundError()
+        }
+
         db.delete(req.params.id)
         res.render('client/deleted')
     })
